refactor: remove legacy codes on perf events
1. remove codes for perf map in `pkg/bpf`

2. remove legacy run_events and replace it with the new one
